Shopping cart (#18)
* added post, get, delete shopping cart

* added patch method for shopping cart

* code style fix

+++ api/utils.py
@@ -9,20 +9,13 @@ from api import filters
 
 
 def check_obj(
-    repository, profile_id, product_id, quantity=None, is_shop=None
-) -> Response:
-    if is_shop is True:
-        if repository.get_obj(profile_id, product_id, quantity).exists():
-            return Response(
-                {'error': 'This product already added'},
-                status=status.HTTP_400_BAD_REQUEST,
-            )
-    else:
-        if repository.get_obj(profile_id, product_id).exists():
-            return Response(
-                {'error': 'This product already added'},
-                status=status.HTTP_400_BAD_REQUEST,
-            )
+    repository, profile_id, product_id
+) -> Response | None:
+    if repository.get_obj(profile_id, product_id).exists():
+        return Response(
+            {'error': 'This product already added'},
+            status=status.HTTP_400_BAD_REQUEST,
+        )
 
 
 def is_shop_check(
@@ -34,25 +27,32 @@ def is_shop_check(
     is_shop: bool = None,
     quantity: int = None,
 ) -> Response:
-    if is_shop is True:
-        check_obj(
-            product_id=product_id,
-            profile_id=profile_id,
-            repository=repository,
-            quantity=quantity,
-            is_shop=is_shop,
-        )
+    response = check_obj(
+        product_id=product_id,
+        profile_id=profile_id,
+        repository=repository
+    )
+    if response:
+        return response
+    if is_shop:
         return repository.create_obj(profile, product, quantity)
     else:
-        check_obj(
-            product_id=product_id,
-            profile_id=profile_id,
-            repository=repository,
-            quantity=quantity,
-            is_shop=is_shop,
-        )
         return repository.create_obj(profile, product)
 
+
+def check_match(
+        request: HttpRequest,
+        profile_id: int,
+        product_id: int
+) -> Response | None:
+    if (
+            profile_id != request.data.get('profile') or
+            product_id != request.data.get('product')
+    ):
+        return Response(
+            {'error': 'Provided data does not match'},
+            status=status.HTTP_400_BAD_REQUEST
+        )
 
 def favorite_or_cart(
     request: HttpRequest,
@@ -69,11 +69,10 @@ def favorite_or_cart(
     product = product_repository.get_obj(product_id)
 
     if request.method == 'POST':
-        if repository.get_obj(profile_id, product_id).exists():
-            return Response(
-                {'error': 'This product already added to favorite'},
-                status=status.HTTP_400_BAD_REQUEST,
-            )
+        if is_shop:
+            quantity = request.data.get('quantity')
+            if response:=check_match(request, profile.id, product.id):
+                return response
         obj = is_shop_check(
             profile=profile,
             product=product,
@@ -83,15 +82,24 @@ def favorite_or_cart(
             is_shop=is_shop,
             quantity=quantity,
         )
+        if isinstance(obj, Response):
+            return obj
         serializer = obj_serializer(obj)
         return Response(serializer.data, status=status.HTTP_201_CREATED)
 
-    if request.method == 'DELETE':
-        if is_shop is True:
-            obj = repository.get_obj(profile_id, product_id, quantity)
-        else:
-            obj = repository.get_obj(profile_id, product_id)
+    if request.method == 'PATCH':
+        quantity = request.data.get('quantity')
+        if response:=check_match(request, profile.id, product.id):
+            return response
+        if (obj := repository.get_obj(profile_id, product_id).first()) is None:
+            return Response(status=status.HTTP_400_BAD_REQUEST)
+        obj.quantity = quantity
+        obj.save()
+        serializer = obj_serializer(obj)
+        return Response(serializer.data, status=status.HTTP_200_OK)
 
+    if request.method == 'DELETE':
+        obj = repository.get_obj(profile_id, product_id)
         if obj.exists():
             obj.delete()
             return Response(status=status.HTTP_204_NO_CONTENT)
